requirements.txt
python-dotenv
caldav
PyMuPDF
Pillow
requests
schedule